refactor: move helper functions to "main_test"
The status listener test file is already pretty big. The goal of moving
as much helper code as possible to the "main_test" file is to reduce
noise and making it easier to reason about what's being tested.
